+++ g_Gemini/source/CLevelDensity.h
@@ -22,50 +22,50 @@ class CLevelDensity
  private:
   CLevelDensity();
   static CLevelDensity *fInstance; //!< instance member to make this class a singleton
-  static float k0; //!< inverse level-density parameter at U=0
-  static float kInfinity; //!< inverse level-density parameter at U=infinity
-  static float aKappa; //!< mass dependence of kappa
-  static float cKappa; //!< mass dependence of kappa
-  static float af_an; //!< ratio of sym-fission saddle  to equilibrium level-density para
-  static float aimf_an; //!< ratio of asy-fission(imf) saddle to equilibrium level-density para
+  static double k0; //!< inverse level-density parameter at U=0
+  static double kInfinity; //!< inverse level-density parameter at U=infinity
+  static double aKappa; //!< mass dependence of kappa
+  static double cKappa; //!< mass dependence of kappa
+  static double af_an; //!< ratio of sym-fission saddle  to equilibrium level-density para
+  static double aimf_an; //!< ratio of asy-fission(imf) saddle to equilibrium level-density para
   static bool normal;
-  static float Ucrit0; //!< Ucrit for J= 0, vanishing of pairing
-  static float Jcrit; //!< spin for the vanishing of pairing
-   float aden; //!< little-density parameter
-  float entropy; //!< entropy
-  float temp;  //!< temperature
-  static float eFade; //!< fade out of shell effects with excitation energy
-  static float jFade; //!< fade out of shell effects with spin
-  float fU;   //!< thermal excitation energy in MeV
-  static float const pi; //!< the mathematical constant \f$\pi\f$
+  static double Ucrit0; //!< Ucrit for J= 0, vanishing of pairing
+  static double Jcrit; //!< spin for the vanishing of pairing
+   double aden; //!< little-density parameter
+  double entropy; //!< entropy
+  double temp;  //!< temperature
+  static double eFade; //!< fade out of shell effects with excitation energy
+  static double jFade; //!< fade out of shell effects with spin
+  double fU;   //!< thermal excitation energy in MeV
+  static double const pi; //!< the mathematical constant \f$\pi\f$
 
  public:
   static CLevelDensity *instance(); //!< instance member to make this class a singleton
-  float getLittleA(int iA,float fU0,float fPairing=0.,float fShell=0.,
-                 float fJ=0.,short iFission=0);
-  float getLittleA(int iA, short iFission);
-  float getU(float fU0,float fPairing, float fShell, float fJ);
-  float getAden();
-  float getLogLevelDensitySpherical(int iA,float fU0,float fPairing,
-	       float fShell,float fJ,float fMinertia,
+  double getLittleA(int iA,double fU0,double fPairing=0.,double fShell=0.,
+                 double fJ=0.,short iFission=0);
+  double getLittleA(int iA, short iFission);
+  double getU(double fU0,double fPairing, double fShell, double fJ);
+  double getAden();
+  double getLogLevelDensitySpherical(int iA,double fU0,double fPairing,
+	       double fShell,double fJ,double fMinertia,
                short iFission=0);
-  float getLogLevelDensitySpherical(int iA,float fU0,float fPairing,
-         float fShell);
-  float getTemp();
-  float getEntropy();
-  static void setLittleA(float k00,float aKappa0=0., float cKappa0=0.,
-          float kInfinity=12.);
-  static void setAfAn(float af_an0);
-  static void setAimfAn(float aimf_an0);
-  static void setUcrit(float Ucrit00, float Jcrit);
-  static float getAKappa();
-  static float getCKappa();
-  static float getK0();
-  static float getKInfinity();
-  static float getAfAn();
-  static float getAimfAn();
+  double getLogLevelDensitySpherical(int iA,double fU0,double fPairing,
+         double fShell);
+  double getTemp();
+  double getEntropy();
+  static void setLittleA(double k00,double aKappa0=0., double cKappa0=0.,
+          double kInfinity=12.);
+  static void setAfAn(double af_an0);
+  static void setAimfAn(double aimf_an0);
+  static void setUcrit(double Ucrit00, double Jcrit);
+  static double getAKappa();
+  static double getCKappa();
+  static double getK0();
+  static double getKInfinity();
+  static double getAfAn();
+  static double getAimfAn();
   static void printParameters();
-  float getLogLevelDensityScission(int iA, float U, float adenInv=8.);
-  float J;
+  double getLogLevelDensityScission(int iA, double U, double adenInv=8.);
+  double J;
 };
 #endif
