+++ g_Gemini/source/CLightP.h
@@ -21,25 +21,25 @@ class CLightP : public CNuclide
   bool mustDelete; //!< bool to remember that we must delete the tLArray
  public:
 
-  CLightP(int iZ,int iA,float fS,string sName);
-  CLightP(int,int,float,CTlBarDist*,CSigBarDist*);
+  CLightP(int iZ,int iA,double fS,string sName);
+  CLightP(int,int,double,CTlBarDist*,CSigBarDist*);
 
    ~CLightP(); //!< destructor
 
    CTlBarDist *tlArray; //!< gives transmission coefficients
    CSigBarDist *sigBarDist; //!< gives inverse xsections for charged part
    
-  float suppress; //!< suppression factor to scale Hauser-Feshbach width
-  float fMInertia; //!< moment of inertia of residue
-  float fMInertiaOrbit; //!< moment of inerria for orbit of light P
-  float separationEnergy; //!< separation energy in MeV
-  float fPair; //!< pairing energy of residue
-  float fShell; //!< shell correction for residue
+  double suppress; //!< suppression factor to scale Hauser-Feshbach width
+  double fMInertia; //!< moment of inertia of residue
+  double fMInertiaOrbit; //!< moment of inerria for orbit of light P
+  double separationEnergy; //!< separation energy in MeV
+  double fPair; //!< pairing energy of residue
+  double fShell; //!< shell correction for residue
   CNuclide residue; //!< contains info on the daughter nucleus  
   bool odd; //!< true for odd mass
   SStoreEvapVector storeEvap; //!<store evaporation info
-  float width; //!<decay width in MeV
-  float rLight; //!< radius of light particle in fm
+  double width; //!<decay width in MeV
+  double rLight; //!< radius of light particle in fm
 };
 
 #endif
