+++ g_Gemini/source/SStoreEvap.h
@@ -15,8 +15,8 @@
 
 struct SStoreEvap
 {
-  float gamma; //!< decay width of channel in MeV
-  float energy; //!< energy of evaporated particle in MeV
+  double gamma; //!< decay width of channel in MeV
+  double energy; //!< energy of evaporated particle in MeV
   short unsigned  spin; //!< spin of daughter
   short unsigned L; //!< orbitla angular momentum of evaporated particle
 };
